Bienvenido

+++ js/main.js
@@ -29,8 +29,18 @@ function agregarAlCarrito(Habitacion){
         Reserva.push(nuevaReserva);
     }
     localStorage.setItem("Reserva", JSON.stringify(Reserva));
-}
+    SeAgregoAlCarrito(Habitacion.nombre)
 
+}
+function SeAgregoAlCarrito(nombre){
+    Swal.fire({
+        position: "top-end",
+        icon: "success",
+        title: "Tu Reserva de " + nombre + " fue confirmada",
+        showConfirmButton: false,
+        timer: 1500
+      });
+}
 
 function CrearCard(Habitacion){
     
@@ -60,28 +70,69 @@ const botonMostrar = document.createElement("button");
 botonMostrar.innerText = "Mostrar Reserva";
 botonMostrar.className = "Boton"
 botonMostrar.addEventListener("click", () => {
-    console.log("Estas son tus Habitaciones",Reserva)
+    MostrarReserva()
 });
 
 main.append(botonMostrar);
+
+function MostrarReserva(){
+    Swal.fire({
+        title: "Reservas Confirmadas",
+        html: Reserva,
+        icon: "info",
+        confirmButtonText: "Correcto"
+    });
+}
+
+
 
 const botonLimpiar = document.createElement("button");
 botonLimpiar.className = "Boton"
 botonLimpiar.innerText = "Cancelar Reservas";
 botonLimpiar.addEventListener("click", () => {
-    Reserva = [];
-    localStorage.setItem("Reserva", JSON.stringify(Reserva))
+    Presionarlimmpiar()
 });
 
+
+function Presionarlimmpiar () {
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+          Reserva = [];
+          localStorage.setItem("Reserva", JSON.stringify(Reserva))
+        }
+      });
+    
+}
 main.append(botonLimpiar);
 
 const botonLimpiar1 = document.createElement("button");
 botonLimpiar1.className = "Boton"
 botonLimpiar1.innerText = "Eliminar Ultima Reserva";
 botonLimpiar1.addEventListener("click", () => {
+    PresionarLimpiar1()
+});
+function PresionarLimpiar1(){
+    Swal.fire({
+        title: "Ultima Reserva Borrada",
+        text: "Satisfactoriamente",
+        icon: "success"
+      });
     Reserva.shift();
     localStorage.setItem("Reserva", JSON.stringify(Reserva))
-});
+}
 
 main.append(botonLimpiar1);
